refactor: 폴더 구조 변경 및 파일 재배치
프로젝트의 가독성 및 유지보수성 향상을 위해 폴더 구조 변경
