+++ pyproject.toml
@@ -14,13 +14,13 @@ dependencies = [
     "google-generativeai>=0.8.5",
     "mcp[cli]>=1.6.0",
     "python-dotenv>=1.1.0",
-    "pynput>=1.7.6",
     "requests",
     "sounddevice>=0.5.1",
     "torch>=2.6.0",
     "torchaudio>=2.6.0",
     "torchvision>=0.21.0",
-    "google-cloud-speech>=2.26.0",
+    "pynput>=1.8.1",
+    "google-cloud-speech>=2.32.0",
 ]
 
 [tool.uv.workspace]
